% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop.R
\name{backprop}
\alias{backprop}
\title{Backpropagation for network with variable layer sizes}
\usage{
backprop(X, y, cache, params)
}
\arguments{
\item{X}{Input matrix}

\item{y}{Target values}

\item{cache}{Forward propagation cache}

\item{params}{Network parameters}
}
\value{
List of gradients
}
\description{
Backpropagation for network with variable layer sizes
}
