% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backprop_namlss.R
\name{backprop_namlss}
\alias{backprop_namlss}
\title{Backpropagation für NAMLSS}
\usage{
backprop_namlss(X, y, fwd, params, dropout_rate = 0)
}
\arguments{
\item{X}{Input Matrix (n_features x batch_size)}

\item{y}{Target Werte}

\item{fwd}{Forward Pass Ergebnisse}

\item{params}{Modellparameter}

\item{dropout_rate}{Dropout Rate}
}
\value{
Liste mit Gradienten
}
\description{
Backpropagation für NAMLSS
}
