% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_feature_net.R
\name{forward_feature_net}
\alias{forward_feature_net}
\title{Forward Pass f端r ein einzelnes Feature-Netzwerk}
\usage{
forward_feature_net(x_j, params, j, dropout_rate = 0, training = FALSE)
}
\arguments{
\item{x_j}{Input f端r Feature j (1 x batch_size)}

\item{params}{Alle Modellparameter}

\item{j}{Feature-Index}

\item{dropout_rate}{Dropout-Rate (nur nach erster Schicht)}

\item{training}{Training-Modus}
}
\value{
Matrix (2 x batch_size) mit Outputs f端r mu und sigma
}
\description{
Forward Pass f端r ein einzelnes Feature-Netzwerk
}
