% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer_dimensions.R
\name{getLayerDimensions}
\alias{getLayerDimensions}
\title{Get Layer Dimensions for a Neural Network}
\usage{
getLayerDimensions(X, out_dim, hidden_neurons, train = TRUE)
}
\arguments{
\item{X}{A numeric matrix or vector representing input data. If a matrix, rows correspond to features and columns to observations.}

\item{out_dim}{Integer specifying the output dimension (e.g., number of output units).}

\item{hidden_neurons}{Integer specifying the number of neurons in the hidden layer.}

\item{train}{Logical value indicating whether the data is used for training (currently not affecting the output). Default is \code{TRUE}.}
}
\value{
A named list with elements:
\item{n_x}{Number of input features (rows of \code{X}).}
\item{n_h}{Number of hidden neurons.}
\item{n_y}{Number of output neurons.}
}
\description{
Determines the input, hidden, and output layer dimensions based on the input matrix
and specified configuration.
}
\examples{
X <- matrix(rnorm(20), nrow = 5)
dims <- getLayerDimensions(X, out_dim = 2, hidden_neurons = 4)
dims$n_x

}
