% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_namls_params.R
\name{init_namls_params}
\alias{init_namls_params}
\title{Initialize NAMLS Parameters}
\usage{
init_namls_params(
  n_features,
  hidden_neurons = c(250, 50, 25),
  y_mean = 1,
  y_sd = 0,
  sigma0 = NULL,
  seed = 42
)
}
\arguments{
\item{n_features}{number of input features (J)}

\item{hidden_neurons}{hidden layer architecture, z.B. c(250, 50, 25)}

\item{y_mean}{mean of target (for numerical stability)}

\item{y_sd}{sd of target (for numerical stability)}

\item{sigma0}{Initial guess for sigma (for numerical stability)}

\item{seed}{seed setting ( default is 42)}
}
\value{
A list containing the initialized parameters for all J sub networks
}
\description{
Initializes NAMLS with J sub networks (one per feature),
each with a two-dimensional output (mu and sigma)
}
