% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_namlss_params.R
\name{init_namlss_params}
\alias{init_namlss_params}
\title{Initialize NAMLSS Parameters}
\usage{
init_namlss_params(
  n_features,
  hidden_neurons = c(250, 50, 25),
  y_mean = 1,
  y_sd = 0,
  sigma0 = NULL,
  seed = 42
)
}
\arguments{
\item{n_features}{Anzahl der Input-Features (J)}

\item{hidden_neurons}{Vector der Hidden Layer Größen, z.B. c(250, 50, 25)}

\item{y_mean}{mean of target (for numerical stability)}

\item{y_sd}{sd of target (for numerical stability)}

\item{sgima0}{Initial guess for sigma (for numerical stability)}
}
\value{
Liste mit initialisierten Parametern für alle J Subnetzwerke
}
\description{
Initialisiert NAMLSS mit J Subnetzwerken (eins pro Feature),
jedes mit 2-dimensionalem Output (für mu und sigma)
}
