% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_arch.R
\name{plot_architecture}
\alias{plot_architecture}
\title{Plots architecture for DNN (NN class)}
\usage{
plot_architecture(model)
}
\arguments{
\item{model}{trained DNN model}
}
\value{
network architecture
}
\description{
Helper function to plot the model architecture after training
}
