% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{pm1_scaler}
\alias{pm1_scaler}
\title{Data Preprocessing Functions
This script contains functions for data preprocessing steps described by
Thielmann et al. (2020), including scaling to \link{-1, 1} and standard normalizing
of target variables.}
\usage{
pm1_scaler(X_train, eps = 1e-08)
}
\description{
Data Preprocessing Functions
This script contains functions for data preprocessing steps described by
Thielmann et al. (2020), including scaling to \link{-1, 1} and standard normalizing
of target variables.
}
