% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_NAMLS.R
\name{predict_namls}
\alias{predict_namls}
\title{Predict method for NAMLS objects}
\usage{
predict_namls(
  object,
  newdata,
  type = c("response", "link"),
  return_contributions = FALSE,
  ...
)
}
\arguments{
\item{object}{A NAMLS object.}

\item{newdata}{Optional data frame with new observations.}

\item{type}{Character, one of \code{c("link","response")}.}

\item{return_contributions}{Logical; if \code{TRUE}, also return contributions.}

\item{...}{Further arguments (ignored), kept for generic consistency.}
}
\value{
A numeric vector or data.frame with predictions.
}
\description{
Predict method for NAMLS objects
}
