% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{scale_minus_one_to_one}
\alias{scale_minus_one_to_one}
\title{Data Preprocessing Functions
This script contains functions for data preprocessing steps described by
Thielmann et al. (2020), including scaling to \link{-1, 1} and standard normalizin
g of target variables.}
\usage{
scale_minus_one_to_one(x)
}
\description{
Data Preprocessing Functions
This script contains functions for data preprocessing steps described by
Thielmann et al. (2020), including scaling to \link{-1, 1} and standard normalizin
g of target variables.
}
