% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.NN}
\alias{summary.NN}
\title{Summary für NN-Objekte}
\usage{
\method{summary}{NN}(
  object,
  data,
  target_col,
  show_plot = TRUE,
  yscale = c("auto", "log", "robust"),
  cap_quantile = 0.99,
  drop_first = 0
)
}
\arguments{
\item{object}{Ein NN-Objekt}

\item{...}{weitere Argumente}
}
\description{
Summary für NN-Objekte
}
